I corrected some errors so that controller bids the correct price based off of the current air temperature both in singe ramp and double ramp mode.
I removed automatic rail responses that were present if the controller lost the bid. The controller now behaves like a passive controller that bids into the market. 

I rearranged the order of operations in double ramp mode so that the controller first resolves any ramp conflicts then responds to the cleared price then determines it's bid price and quantity.

git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@1773 54b63ff4-7b51-0410-8bda-9fa96797e505