changed generated C source code to #include "plc.h" instead of <plc.h>.
Made compilation more like how it is done in core/load.c (the C compiler
defaults to the value of CC, CFLAGS, if defined, is passed to the compile
step, and LDFLAGS, if defined, is passed to the link step).

Generated files are removed after they are no longer needed.



git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@2382 54b63ff4-7b51-0410-8bda-9fa96797e505